chore: migrate metadata upload
