+++ apps/backend/.dev.vars.example
@@ -3,3 +3,4 @@ EVER_ACCESS_SECRET=
 LOGTAIL_API_KEY=
 INGEST_REST_ENDPOINT=
 IP_API_KEY=
+WALLET_PRIVATE_KEY= #pk without 0x
