+++ apps/backend/src/helpers/constants.ts
@@ -1,3 +1,5 @@
 export const LENS_API_URL = 'https://api-v2.lens.dev/'
 export const PUBLIC_ETHEREUM_NODE = 'https://ethereum.publicnode.com'
+export const IRYS_NODE_URL = 'http://node2.irys.xyz/tx/matic'
+
 export const ERROR_MESSAGE = 'Something went wrong'
