+++ apps/backend/src/index.ts
@@ -1,6 +1,7 @@
 import { Hono } from 'hono'
 
 import did from './routes/did'
+import metadata from './routes/metadata'
 import recommendations from './routes/recommendations'
 import sts from './routes/sts'
 import tail from './routes/tail'
@@ -14,6 +15,7 @@ app.route('/sts', sts)
 app.route('/tail', tail)
 app.route('/tower', tower)
 app.route('/did', did)
+app.route('/metadata', metadata)
 app.route('/recommendations', recommendations)
 
 export default app
