--- packages/workers/metadata/.dev.vars.example
@@ -1,2 +1,0 @@
-# pk without 0x
-WALLET_PRIVATE_KEY=""
