--- packages/workers/metadata/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@workers/metadata",
-  "version": "0.0.0",
-  "license": "AGPL-3.0",
-  "private": true,
-  "scripts": {
-    "worker:dev": "wrangler dev",
-    "worker:deploy": "wrangler deploy"
-  },
-  "dependencies": {
-    "base64url": "^3.0.1",
-    "viem": "^1.21.4"
-  },
-  "devDependencies": {
-    "@tape.xyz/config": "*",
-    "typescript": "^5.3.3"
-  }
-}
