--- packages/workers/metadata/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "@tape.xyz/config/base.tsconfig.json",
-  "compilerOptions": {
-    "lib": ["es2021"],
-    "noEmit": true,
-    "types": ["@cloudflare/workers-types"]
-  }
-}
