--- packages/workers/metadata/wrangler.toml
@@ -1,11 +1,0 @@
-name = "metadata"
-main = "src/index.ts"
-compatibility_date = "2023-01-25"
-routes = [
-	{ pattern = "metadata.tape.xyz", custom_domain = true }
-]
-keep_vars = true
-node_compat = true
-
-[env.production.vars]
-WALLET_PRIVATE_KEY = ""
